requirements.txt
# AI Prediction Battle — pinned dependencies
# Last updated: 2025-04
# Pin exact versions for reproducible installs.

# Environment
python-dotenv==1.0.1

# HTTP
requests==2.32.3

# Data validation
pydantic==2.7.1

# LLM providers
openai==1.30.1
google-generativeai==0.7.2
anthropic==0.28.0

# Web research
tavily-python==0.3.3

# CLI / terminal UI
rich==13.7.1

# Voice (V2)
edge-tts==6.1.12
pygame==2.5.2

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-mock==3.14.0

# Code quality
black==24.4.2
flake8==7.0.0
